Denormalize wordCount field
